Remove matched addresses from sync.
Google Groups can automatically change the email address formatting so added a check which compares insertable and removable addresses in lowercase without dots. This ensures that the sync tool does not try to add or remove addresses that should not. This can break the tool in some cases but for now is a good-enough solution.